+++ semana12/labeEddit/src/App.js
@@ -1,10 +1,19 @@
 import React from 'react'
+import Router from './router/Router'
+import { ChakraProvider, Box } from "@chakra-ui/core"
+import { BrowserRouter } from 'react-router-dom'
+
+
 
 const App = () => {
   return (
-    <div>
-      
-    </div>
+    <ChakraProvider>
+      <BrowserRouter>
+        <Box bg="blue.400" h="70px" w="100%" marginBottom="20" >
+        </Box>
+        <Router />
+      </BrowserRouter>
+    </ChakraProvider>
   )
 }
