--- semana12/labeEddit/src/screens/LoginPage/LoginPage.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const LoginPage = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default LoginPage
