--- semana12/labeEddit/src/screens/PostPage/PostPage.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const PostPage = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default PostPage
