+++ semana12/labeEddit/src/screens/SingUpPage/SingUpPage.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const SingUpPage = () => {
+  return (
+    <div>
+      SingUpPage
+    </div>
+  )
+}
+
+export default SingUpPage
